Set the fillfactor on a number of puppetdb indexes
This commit lowers the fillfactor on a number of indexes that have
shown to get fragmented and have reduced leaf page density from
pgstatindex output.

On tables that receive a lot of updates it is better to leave some
free space in each page to allow the updates to fit in the same
page instead of needing to split the page to store the update.
Setting the fillfactor lower is what leaves the free space.