Restrict indexes settings to 2016.4
Prior to this commit, the index_settings would attempt to be
applied on all versions of PE.  Since the index names may have
changed or maybe the indexes didn't even exist in a past version
this could cause problems.

After this commit, the index_settings will only apply on 2016.4
for now and future iterations could allow figuring out which
indexes exist in different versions of PE and probably tuning them.